Fixed fetching of index
Previously was using sheetId which could be non uniform. I.e. adding and
removing sheets could cause the numbers to be inconsistent with what
ends up in the archive.
Using the index in which they appear in the sheetsElements allows for
this inconsistency, this way the first sheet is always index 1, and is
sheet1.xml in the archive.

Fixes #4